Athlete_Two.java
I uploaded a second copy of Athlete with somee changes/implementations of the required implementations. I worked over it in the weekend and I'm sorry I left out the comments. I had to really plan this out due to the what I had to use was not the best way of going about it. The file needs cleaning up but I'll work on that later. What should be noted is it prints the athletes using Stacks by the lowest score.