Why don't we use the method fillItUp?
I feel like we should use it because since it creates an array of athletes we are able to use the array indexes to compare the times?  I debugged the code and the compare method in the arrays.sort isn't even implemented.  I've also ran the code quite a few times and it has not put any of the athletes in order of time.