Updated the Code
Pretty much everything is finished except for printing the data. The data is in several Maps and there is a comparator that makes sure the data is in order - or it should make sure the data is in order. However, I am stuck: 

The totalTimes are in a map and they are sorted but I need the keys in the sorted corresponding order. For it to mean anything. It's kinda hard to explain this over text so the next time we meet I will give a more thorough explanation.