Figured it out.
The problem I had was after the total times have been sorted and I needed to find the keys of the Map in the same sorted way. So I assigned the keyset of the athlete's total time to a Set. I iterate through that Set and I typecast the elements in the Set into a integer array. Now I have the keys in order and I can print all the athletes.  I did just that it prints the athlete's corresponding information but the only flaw is that the comparator is still wonky. It doesn't always sort.