Map Demo
The code I used to sort the Map. The only difference between this and the contents of the Athtlete class is that the value type for the map is a String rather than a Double.

I found out why it does not always sort when using Strings. When comparing Strings it compares it lexicographcally - Look at the compareTo Method in the String class - meaning that it compares the unicode values and not the value of the String itself. So when the comparator sorted the values it was sorting it based on the lexicographical order. To have it displayed in the correct order the value type of the Map needed to change from String to Double then it would sort properly.