Create UpdatedAthleteWithDatabase
Using Kun's Athlete class, I connected to the database and added the generated information and using a gui, We are able to see the results of the data.  I used my previous classes for database as the driver and the controller. I will show you guys tomorrow morning.  I also need to add a top male and female and also add the ranks! See everyone later!